refine liquidnote-script moves content to trigger
New liquidnote-script move the trigger_container to the trigger link on
show (toggle), and moves it to the end of the body on hide. Also the
appendto and the insertAfter stop the media from playing - nice plus -
so I remove the iframe src reset trick. Code is smaller and more
flexible.